Moved gross rent multiplier component code to own directory
