--- src/pages/GrossRentMultiplier.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-export const GrossRentMultiplier = () => {
-    return (
-        <div>
-            <h1 className="page-title">Gross Rent Multiplier</h1>
-            <p>Compare gross rent multipliers by city or zip code.</p>
-        </div>
-    )
-}
